+++ docs/changelog.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Added SelectedDocument model with validations for assignment, google_doc_id, title, and url.
 - Added "Just finished" to home page.
 - Added Rubric model with validations for assignment.
+- Added Criterion model with validations for rubric, title, description, and position.
 
 ### Removed
 - Removed all classes from `application.html.erb` and `home/index.html.erb`, created backups of each.